implemented Result pattern, refactored code to use it

+++ ReqSense.Application/Services/RequirementService.cs
@@ -2,7 +2,7 @@
 using Microsoft.EntityFrameworkCore;
 using ReqSense.Application.Common.DTOs.Requirement.Request;
 using ReqSense.Application.Common.DTOs.Requirement.Response;
-using ReqSense.Application.Common.Exceptions;
+using ReqSense.Application.Common.Errors;
 using ReqSense.Application.Common.Interfaces;
 using ReqSense.Domain.Entities;
 
@@ -25,39 +25,61 @@ public class RequirementService(
         return mapper.Map<IEnumerable<RequirementBriefDto>>(requirements);
     }
 
-    public async Task<RequirementFullDto> GetRequirementByIdAsync(long requirementId)
+    public async Task<Result<RequirementFullDto>> GetRequirementByIdAsync(long requirementId)
     {
         var requirement = await dbContext.Requirements
             .Include(e => e.Creator)
             .Include(e => e.LastEditor)
-            .SingleAsync(e => e.Id.Equals(requirementId));
+            .FirstOrDefaultAsync(e => e.Id.Equals(requirementId));
 
-        return mapper.Map<RequirementFullDto>(requirement);
+        if (requirement is null)
+        {
+            return Result.Fail<RequirementFullDto>(RequirementErrors.NotFound(requirementId));
+        }
+
+        return Result.Ok(mapper.Map<RequirementFullDto>(requirement));
     }
 
-    public async Task<long> CreateRequirementAsync(CreateRequirementDto dto, long projectId)
+    public async Task<Result<long>> CreateRequirementAsync(CreateRequirementDto dto, long projectId)
     {
+        var existingRequirement = await dbContext.Requirements.FirstOrDefaultAsync(p =>
+            p.Title.Equals(dto.Title) && p.ProjectId.Equals(projectId));
+        if (existingRequirement is not null)
+        {
+            return Result.Fail<long>(RequirementErrors.DuplicateTitle(dto.Title));
+        }
+
         var requirement = mapper.Map<Requirement>(dto);
         requirement.ProjectId = projectId;
 
         dbContext.Requirements.Add(requirement);
         await dbContext.SaveChangesAsync();
-        return requirement.Id;
+        return Result.Ok(requirement.Id);
     }
 
-    public async Task UpdateRequirementAsync(UpdateRequirementDto dto)
+    public async Task<Result> UpdateRequirementAsync(UpdateRequirementDto dto)
     {
         var requirement = await dbContext.Requirements.FindAsync(dto.Id);
-        NotFoundException.ThrowIfNull(requirement, $"Requirement with id {dto.Id} was not found!");
+        if (requirement is null)
+        {
+            return Result.Fail(RequirementErrors.NotFound(dto.Id));
+        }
+
         mapper.Map(dto, requirement);
         await dbContext.SaveChangesAsync();
+        return Result.Ok();
     }
 
-    public async Task DeleteRequirementAsync(long requirementId)
+    public async Task<Result> DeleteRequirementAsync(long requirementId)
     {
         var requirement = await dbContext.Requirements.FindAsync(requirementId);
-        NotFoundException.ThrowIfNull(requirement, $"Requirement with id {requirementId} was not found");
+        if (requirement is null)
+        {
+            return Result.Fail(RequirementErrors.NotFound(requirementId));
+        }
+
         dbContext.Requirements.Remove(requirement!);
         await dbContext.SaveChangesAsync();
+        return Result.Ok();
     }
 }
